+++ packages/app/src/components/catalog/EntityPage.tsx
@@ -25,7 +25,6 @@ import {
   EntityMembersListCard,
   EntityOwnershipCard,
 } from '@backstage/plugin-org';
-import { EntityTechdocsContent } from '@backstage/plugin-techdocs';
 import { EmptyState } from '@backstage/core-components';
 import {
   Direction,
@@ -48,8 +47,10 @@ import {
   RELATION_PROVIDES_API,
 } from '@backstage/catalog-model';
 
-import { TechDocsAddons } from '@backstage/plugin-techdocs-react';
-import { ReportIssue } from '@backstage/plugin-techdocs-module-addons-contrib';
+import {
+  EntityGrafanaDashboardsCard,
+  isDashboardSelectorAvailable,
+} from '@k-phoen/backstage-plugin-grafana';
 
 function isLinksAvailable(entity: Entity) {
   if (entity?.metadata?.links?.length) {
@@ -57,14 +58,6 @@ function isLinksAvailable(entity: Entity) {
   }
   return false;
 };
-
-const techdocsContent = (
-  <EntityTechdocsContent>
-    <TechDocsAddons>
-      <ReportIssue />
-    </TechDocsAddons>
-  </EntityTechdocsContent>
-);
 
 const circleCIContent = (
   <EntitySwitch>
@@ -129,6 +122,7 @@ const overviewContent = (
     <Grid item md={6} xs={12}>
       <EntityCatalogGraphCard variant="gridItem" height={400} />
     </Grid>
+    
     <EntitySwitch>
       <EntitySwitch.Case if={isLinksAvailable}>
         <Grid item md={4} xs={12}>
@@ -143,10 +137,6 @@ const baseEntityPage = (
   <EntityLayout>
     <EntityLayout.Route path="/" title="Overview">
       {overviewContent}
-    </EntityLayout.Route>
-
-    <EntityLayout.Route path="/docs" title="Docs">
-      {techdocsContent}
     </EntityLayout.Route>
   </EntityLayout>
 );
@@ -166,10 +156,6 @@ const appcatalogEntityPage = (
           <EntityDependsOnResourcesCard variant="gridItem" />
         </Grid>
       </Grid>
-    </EntityLayout.Route>
-
-    <EntityLayout.Route path="/docs" title="Docs">
-      {techdocsContent}
     </EntityLayout.Route>
   </EntityLayout>
 );
@@ -210,9 +196,6 @@ const defaultEntityPage = (
       </Grid>
     </EntityLayout.Route>
 
-    <EntityLayout.Route path="/docs" title="Docs">
-      {techdocsContent}
-    </EntityLayout.Route>
   </EntityLayout>
 );
 
@@ -268,16 +251,28 @@ const groupPage = (
   <EntityLayout>
     <EntityLayout.Route path="/" title="Overview">
       <Grid container spacing={3}>
+
         {entityWarningContent}
+
         <Grid item xs={12} md={6}>
           <EntityGroupProfileCard variant="gridItem" />
         </Grid>
         <Grid item xs={12} md={6}>
           <EntityOwnershipCard variant="gridItem" />
         </Grid>
+        
         <Grid item xs={12}>
           <EntityMembersListCard />
         </Grid>
+
+        <EntitySwitch>
+          <EntitySwitch.Case if={e => !!isDashboardSelectorAvailable(e)}>
+            <Grid item xs={12}>
+              <EntityGrafanaDashboardsCard title="Grafana Cloud dashboards" />
+            </Grid>
+          </EntitySwitch.Case>
+        </EntitySwitch>
+
       </Grid>
     </EntityLayout.Route>
   </EntityLayout>