+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Change condition for when to show links. Now links can be shown for all entity kinds.
+- Remove docs tab from entity page.
+- Add list of Grafana dashboards owned by the team to Group entity page.
 
 ## [0.3.0] - 2023-08-18